requirements.txt
django
djangorestframework
djangorestframework-simplejwt
pillow
python-dotenv
google-genai
django-cors-headers
